+++ src/components/vertical-menu/vertical-menu.pw.tsx
@@ -20,6 +20,7 @@ import {
   checkGoldenOutline,
   assertCssValueIsApproximately,
   checkAccessibility,
+  waitForAnimationEnd,
 } from "../../../playwright/support/helper";
 import {
   verticalMenuComponent,
@@ -314,6 +315,42 @@ test.describe("with beforeEach for VerticalMenuFullScreen", () => {
     await mount(<VerticalMenuFullScreenCustom isOpen />);
 
     await expect(verticalMenuItem(page).first()).toBeVisible();
+  });
+
+  test("when a Vertical Menu Fullscreen is opened and then closed, the call to action element should be focused", async ({
+    mount,
+    page,
+  }) => {
+    await mount(<VerticalMenuFullScreenCustom />);
+
+    const item = page.getByRole("button").filter({ hasText: "Menu" });
+    await item.click();
+    await waitForAnimationEnd(verticalMenuFullScreen(page));
+    const closeButton = page.getByLabel("Close");
+    await closeButton.click();
+    await expect(item).toBeFocused();
+  });
+
+  test("when Vertical Menu Fullscreen is open on render, then closed, opened and then closed again, the call to action element should be focused", async ({
+    mount,
+    page,
+  }) => {
+    await mount(<VerticalMenuFullScreenCustom isOpen />);
+
+    await waitForAnimationEnd(verticalMenuFullScreen(page));
+    await expect(verticalMenuFullScreen(page)).toBeVisible();
+    const closeButton = page.getByLabel("Close");
+    await closeButton.click();
+
+    const item = page.getByRole("button").filter({ hasText: "Menu" });
+    await expect(item).not.toBeFocused();
+    await expect(verticalMenuFullScreen(page)).not.toBeVisible();
+
+    await item.click();
+    await waitForAnimationEnd(verticalMenuFullScreen(page));
+    await expect(verticalMenuFullScreen(page)).toBeVisible();
+    await closeButton.click();
+    await expect(item).toBeFocused();
   });
 
   test(`should verify that Vertical Menu Fullscreen has no effect on the tab order when isOpen prop is false`, async ({